requirements.txt
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
packaging==24.2
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0
wsproto==1.2.0
